move function to query
