+++ packages/toolkit/src/lib/react-query-service/knowledge/useCreateKnowledgeBase.ts
