+++ packages/toolkit/src/lib/react-query-service/knowledge/useDeleteKnowledgeBase.ts
